requirements.txt
python-dotenv>=1.0.0
httpx>=0.25.0
fastapi>=0.100.0
uvicorn>=0.23.0
